break out gazebo_drone_pose and wrap in class so can googletest with UDPSender mocked
