--- pose/gazebo/src/send_drone_pose_pi.hpp
@@ -1,8 +1,0 @@
-#include <gazebo/msgs/msgs.hh>
-
-#ifndef SEND_DRONE_POSE_PI
-#define SEND_DRONE_POSE_PI
-
-void cbLocalPose(ConstPosesStampedPtr& msg);
-
-#endif
