+++ pose/src/udp_sender.hpp
@@ -12,8 +12,8 @@
 class UDPSender {       // The class
     public:             // Access specifier
         UDPSender(std::string host, unsigned short int dest_port);  // Constructor
-        ~UDPSender();
-        void send_pose_message(const PoseTransfer::PoseMessage pose_message);  // Method/function declaration
+        virtual ~UDPSender();
+        virtual void send_pose_message(const PoseTransfer::PoseMessage pose_message);  // Method/function declaration
     private:
         unsigned short int dest_port;
         std::string host;  // Attribute (string variable)
